Split AuthRoutesTest into indiviual test classes
